pyproject.toml
[project]
name = "alliance-label-generator"
version = "0.1.0"
description = "Print-ready PDF label generator for chemical products"
requires-python = ">=3.11"
dependencies = [
    "reportlab>=4.0",
    "python-barcode>=0.15",
    "qrcode>=7.4",
    "svglib>=1.5",
    "pydantic>=2.0",
    "typer>=0.9",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pdf2image>=1.16",
]

[project.scripts]
label = "src.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
